cmd/snapd/tool/snap-gpio-helper: move snap-gpio-helper sources around
Move snap-gpio-helper sources around before transitioning to multi
entry dispatch mechanism.

Related: SNAPDENG-37215
