+++ cmd/snap-gpio-helper/main.go
@@ -20,46 +20,9 @@
 package main
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/jessevdk/go-flags"
-
-	"github.com/snapcore/snapd/sandbox/gpio"
-	"github.com/snapcore/snapd/snapdtool"
+	"github.com/snapcore/snapd/cmd/snapd/tool/snap-gpio-helper"
 )
 
-type options struct {
-	CmdExportChardev   cmdExportChardev   `command:"export-chardev"`
-	CmdUnexportChardev cmdUnexportChardev `command:"unexport-chardev"`
-}
-
-var gpioEnsureAggregatorDriver = gpio.EnsureAggregatorDriver
-
-func run(args []string) error {
-	// Make sure the gpio-aggregator module is loaded because the
-	// systemd security backend comes before the kmod security
-	// backend, there is an edge case on first connection where
-	// the helper service could be started before the gpio-aggregator
-	// module is loaded.
-	if err := gpioEnsureAggregatorDriver(); err != nil {
-		return nil
-	}
-
-	var opts options
-	p := flags.NewParser(&opts, flags.HelpFlag|flags.PassDoubleDash)
-
-	if _, err := p.ParseArgs(args); err != nil {
-		return err
-	}
-	return nil
-}
-
 func main() {
-	snapdtool.ExecInSnapdOrCoreSnap()
-
-	if err := run(os.Args[1:]); err != nil {
-		fmt.Fprintf(os.Stderr, "error: %v\n", err)
-		os.Exit(1)
-	}
+	snap_gpio_helper.Main()
 }
