+++ cmd/snapd/tool/snap-gpio-helper/cmd_export.go
@@ -17,7 +17,7 @@
  *
  */
 
-package main
+package snap_gpio_helper
 
 import (
 	"context"
